add function call conventions
the caller puts the function parameters onto the stack in inverted order and calls the function using the call instruction; the callee first pushes the base pointer onto the stack and then copies the value of the stack pointer into the base pointer, then allocates space on stack for local variables if necessary and then pushed onto the stack the registers that will be used in the function. The aprameters can be accessed starting from %bp+8 (where the first parameter is located). Both the caller and callee must restore the stack back to its original state