create exit syscall and fix program execution
now, when a program is executed, it goes through the scheduler and is no longer executed in the interrupt; one of the previous problems was that a user space process was executed while still in the keyboard interrupt handler. As the user space process returns by performing the exit syscall, the program execution wasn't continued from where it left (this means that the irq handler wasn't finished, so it didn't tell the PIC for example that the interrupt was handled). Now the execution flow is the folowing: when a file is executed, a task is created and put in the scheduler's task queue, allowing the keyboard irq to finish normally. When the scheduler looks in the queue, it will find and execute the task.