fix some stack problems
there were two major problems regarding the stack. Problem 1: when a kernel task is interrupted, the CPU stores the EFLAGS, CS and EIP to the stack before entering the irq/isr handler. The problem appears when another task is scheduled to run in ring 3, in which case the IRET instruction before leaving the handler takes from the stack EIP, CS, EFLAGS, USERESP and SS. The last two registers were not stored be the CPU and my previous implmentation included them on the stack, overwriting other data present on the stack. Problem 2: when a task from user space is interrupted, the CPU pushes SS, ESP, EFLAGS, CS, EIP to the stack and enters the isr/irq handler. When another task from kernel space is scheduled, the IRET instruction will only take the EIP, CS and EFLAGS from the stack, leaving ESP and SS untouched.